refactor: storage related models are in base storage client
