+++ src/crawlee/__init__.py
@@ -1,6 +1,6 @@
 from importlib import metadata
 
-from ._models import Request
+from ._request import Request
 from ._types import ConcurrencySettings, EnqueueStrategy
 from ._utils.globs import Glob
