+++ src/crawlee/base_storage_client/_base_dataset_collection_client.py
@@ -4,7 +4,7 @@ from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from crawlee._models import DatasetListPage, DatasetMetadata
+    from crawlee.base_storage_client._models import DatasetListPage, DatasetMetadata
 
 
 class BaseDatasetCollectionClient(ABC):
