+++ src/crawlee/base_storage_client/_base_key_value_store_client.py
@@ -6,7 +6,11 @@ from typing import TYPE_CHECKING, Any, AsyncContextManager
 if TYPE_CHECKING:
     from httpx import Response
 
-    from crawlee._models import KeyValueStoreListKeysPage, KeyValueStoreMetadata, KeyValueStoreRecord
+    from crawlee.base_storage_client._models import (
+        KeyValueStoreListKeysPage,
+        KeyValueStoreMetadata,
+        KeyValueStoreRecord,
+    )
 
 
 class BaseKeyValueStoreClient(ABC):
