+++ src/crawlee/memory_storage_client/_key_value_store_client.py
@@ -10,18 +10,18 @@ from typing import TYPE_CHECKING, Any, AsyncContextManager
 
 from typing_extensions import override
 
-from crawlee._models import (
+from crawlee._types import StorageTypes
+from crawlee._utils.crypto import crypto_random_object_id
+from crawlee._utils.data_processing import maybe_parse_body, raise_on_duplicate_storage, raise_on_non_existing_storage
+from crawlee._utils.file import determine_file_extension, force_remove, force_rename, is_file_or_bytes, json_dumps
+from crawlee.base_storage_client import BaseKeyValueStoreClient
+from crawlee.base_storage_client._models import (
     KeyValueStoreKeyInfo,
     KeyValueStoreListKeysPage,
     KeyValueStoreMetadata,
     KeyValueStoreRecord,
     KeyValueStoreRecordMetadata,
 )
-from crawlee._types import StorageTypes
-from crawlee._utils.crypto import crypto_random_object_id
-from crawlee._utils.data_processing import maybe_parse_body, raise_on_duplicate_storage, raise_on_non_existing_storage
-from crawlee._utils.file import determine_file_extension, force_remove, force_rename, is_file_or_bytes, json_dumps
-from crawlee.base_storage_client import BaseKeyValueStoreClient
 from crawlee.memory_storage_client._creation_management import (
     find_or_create_client_by_id_or_name_inner,
     persist_metadata_if_enabled,
