+++ src/crawlee/memory_storage_client/_request_queue_collection_client.py
@@ -4,8 +4,8 @@ from typing import TYPE_CHECKING
 
 from typing_extensions import override
 
-from crawlee._models import RequestQueueListPage, RequestQueueMetadata
 from crawlee.base_storage_client import BaseRequestQueueCollectionClient
+from crawlee.base_storage_client._models import RequestQueueListPage, RequestQueueMetadata
 from crawlee.memory_storage_client._creation_management import get_or_create_inner
 from crawlee.memory_storage_client._request_queue_client import RequestQueueClient
