+++ src/crawlee/sessions/_models.py
@@ -1,4 +1,4 @@
-# ruff: noqa: TCH002 TCH003
+# ruff: noqa: TCH001, TCH002, TCH003 (because of Pydantic)
 
 from __future__ import annotations
