+++ src/crawlee/storages/_key_value_store.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING, Any, AsyncIterator, TypeVar, overload
 
 from typing_extensions import override
 
-from crawlee._models import KeyValueStoreKeyInfo, KeyValueStoreMetadata
+from crawlee.base_storage_client._models import KeyValueStoreKeyInfo, KeyValueStoreMetadata
 from crawlee.storages._base_storage import BaseStorage
 
 if TYPE_CHECKING:
