+++ src/crawlee/storages/_request_queue.py
@@ -9,16 +9,11 @@ from typing import TYPE_CHECKING, Any, Generic, TypedDict, TypeVar
 
 from typing_extensions import override
 
-from crawlee._models import (
-    BaseRequestData,
-    ProcessedRequest,
-    Request,
-    RequestQueueMetadata,
-)
 from crawlee._utils.crypto import crypto_random_object_id
 from crawlee._utils.lru_cache import LRUCache
 from crawlee._utils.requests import unique_key_to_request_id
 from crawlee._utils.wait import wait_for_all_tasks_for_finish
+from crawlee.base_storage_client._models import ProcessedRequest, RequestQueueMetadata
 from crawlee.events._types import Event
 from crawlee.storages._base_storage import BaseStorage
 from crawlee.storages._request_provider import RequestProvider
@@ -26,9 +21,10 @@ from crawlee.storages._request_provider import RequestProvider
 if TYPE_CHECKING:
     from collections.abc import Sequence
 
+    from crawlee._request import BaseRequestData, Request
     from crawlee.base_storage_client import BaseStorageClient
     from crawlee.configuration import Configuration
-    from crawlee.events._event_manager import EventManager
+    from crawlee.events import EventManager
 
 logger = getLogger(__name__)
